Add transformer for defendant
Defendant models for DocGen have been added, and the SDK version has
also been changed in the Prosecution POM. As a result of this, a unit
test for the ProsecutionCase transformer also had to be changed. The
defendant mapper has also been added for transforming the SDK models
to those in DocGen.

Resolves SJP-692